Add some docs outlining potential classes/properties
There are 22 files in the Misc folder. Each represents an Aras package in an OOTB 11 SP9 database. (Four OOTB packages do not contain ItemTypes, so they are skipped)

Not every "class" or "property" in these files will become part of the base ontology. I generated these simply to take an inventory of Aras ItemTypes and their properties.