Fix properties with >1 domain or range
Previously, properties with multiple domains or ranges were defined for an intersection of those classes. Ex: the file_type property would only apply to individuals that were both a CADFiles item and File item.

Now these property definitions wrap the domains and/or ranges in a unionOf so they will apply to any individual from a listed class.